+++ app/Models/Item.php
@@ -121,9 +121,9 @@ class Item extends BaseModel
     {
         return $this->morphToMany(
             AbilityScore::class,
-            'entity',
+            'reference',
             'entity_saving_throws',
-            'entity_id',
+            'reference_id',
             'ability_score_id'
         )
             ->withPivot('dc', 'save_effect', 'is_initial_save', 'save_modifier')